Fix HID stream buffering
Unread HID reports get buffered until they're read, so if we fail to
call Read as often as the device emits reports, we will accrue stale
reports in the buffer and fall further and further behind. This can be
seen when running CPU intensive tasks that starve Fan Control of cycles,
or induced manually by using Resource Monitor to suspend the process.

Fix this by spawning a thread that spins on HidStream.Read to ensure
that we're always emptying the HID stream's buffer.